requirements.txt
flask>=3.0
flask-socketio>=5.3
flask-cors>=4.0
python-socketio>=5.10
eventlet>=0.36
PyJWT>=2.8
python-dotenv>=1.0
claude-code-sdk>=0.1
